Update README.md
In machine learning, supervised learning methods for binary classification are algorithms that categorize new observations into one of two classes. This work aims to evaluate the performance of three supervised learning approaches on a binary classification problem. It aims to devise a better understanding of both the benefits and limitations of each classification method employed to classify a dataset containing data of two classes: class 0, and class 1. The three binary classification techniques implemented in the course of this work include K nearest Neighbour, Decision Tree, and Logistic Regression.